[feature] Set up link between log in and sign up activities
