+++ app/src/main/java/com/andela/philskiiiwalker/brighteventsandroidapp/home/MainActivity.java
@@ -1,13 +1,10 @@
 package com.andela.philskiiiwalker.brighteventsandroidapp.home;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 
 import com.andela.philskiiiwalker.brighteventsandroidapp.R;
-import com.andela.philskiiiwalker.brighteventsandroidapp.registration.SignUpActivity;
 
 public class MainActivity extends AppCompatActivity {
     Button getStarted;
@@ -17,14 +14,5 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         getStarted = findViewById(R.id.signUpButton);
-
-        getStarted.setOnClickListener(new Button.OnClickListener(){
-
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this, SignUpActivity.class);
-                startActivity(intent);
-            }
-        });
     }
 }
