--- app/src/main/java/com/andela/philskiiiwalker/brighteventsandroidapp/registration/SignUpActivity.java
@@ -1,15 +1,0 @@
-package com.andela.philskiiiwalker.brighteventsandroidapp.registration;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
-import com.andela.philskiiiwalker.brighteventsandroidapp.R;
-
-public class SignUpActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_signup);
-    }
-}
