+++ app/src/main/java/com/andela/philskiiiwalker/brighteventsandroidapp/registration/service/BrightEventsAPI.java
@@ -1,0 +1,9 @@
+package com.andela.philskiiiwalker.brighteventsandroidapp.registration.service;
+
+import retrofit2.Call;
+import retrofit2.http.POST;
+
+//public interface BrightEventsAPI {
+//    @POST("https://bright-events-api-.herokuapp.com/api/v2/auth/register")
+//    Call<RegisterUserResponse> getResults();
+//}
