+++ app/src/main/java/com/andela/philskiiiwalker/brighteventsandroidapp/splash_screen/view/SplashScreen.java
@@ -6,6 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 
 import com.andela.philskiiiwalker.brighteventsandroidapp.R;
 import com.andela.philskiiiwalker.brighteventsandroidapp.home.MainActivity;
+import com.andela.philskiiiwalker.brighteventsandroidapp.login.LoginActivity;
 
 public class SplashScreen extends AppCompatActivity {
     @Override
@@ -18,7 +19,7 @@ public class SplashScreen extends AppCompatActivity {
             public void run() {
                 try {
                     sleep(3000);
-                    Intent intent = new Intent(getApplicationContext(), MainActivity.class);
+                    Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
                     startActivity(intent);
                     finish();
                 } catch (InterruptedException e) {
