+++ app/src/main/res/layout/activity_signin.xml
@@ -1,28 +1,26 @@
-<RelativeLayout android:layout_width="match_parent"
+<RelativeLayout xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android">
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <android.support.constraint.ConstraintLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true">
 
         <TextView
             android:id="@+id/signinTitleTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_alignParentLeft="true"
             android:layout_alignParentStart="true"
             android:layout_alignParentTop="true"
+            android:layout_marginLeft="23dp"
             android:layout_marginStart="23dp"
             android:layout_marginTop="18dp"
             android:text="@string/sign_in_title"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="23dp" />
-
-        <EditText
-            android:id="@+id/signinEmailEditText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:ems="10"
-            android:inputType="textEmailAddress"
-            android:text="@string/email" />
+            android:textSize="30sp" />
 
         <EditText
             android:id="@+id/signinUsernameEditText"
@@ -30,48 +28,51 @@
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginBottom="169dp"
             android:ems="10"
+            android:hint="@string/username"
             android:inputType="textPersonName"
-            android:text="@string/username" />
+            tools:layout_editor_absoluteY="86dp"
+            app:layout_constraintTop_toBottomOf="@id/signinTitleTextView"/>
 
         <EditText
-            android:id="@+id/editText"
+            android:id="@+id/signinPasswordEditText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
             android:layout_marginBottom="104dp"
             android:ems="10"
+            android:hint="@string/password"
             android:inputType="textPassword"
-            android:text="@string/password"/>
+            app:layout_constraintTop_toBottomOf="@id/signinUsernameEditText" />
 
-        <TextView
-            android:id="@+id/signinEmailTextView"
+        <Button
+            android:id="@+id/signInNowButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_above="@+id/signinEmailEditText"
-            android:layout_alignStart="@+id/signinEmailEditText"
-            android:text="@string/email_title"
-            android:layout_alignLeft="@+id/signinEmailEditText" />
+            android:layout_alignParentBottom="true"
+            android:layout_centerHorizontal="true"
+            android:layout_marginBottom="40dp"
+            android:text="@string/sign_in_title"
+            app:layout_constraintTop_toBottomOf="@id/signinPasswordEditText" />
 
         <TextView
-            android:id="@+id/signinUsernameTextView"
+            android:id="@+id/signInToSignUpInstructionTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/signinEmailEditText"
-            android:layout_below="@+id/signinEmailEditText"
-            android:text="@string/username"
-            android:layout_alignLeft="@+id/signinEmailEditText" />
-
+            android:text="@string/signin_to_signup_textview"
+            app:layout_constraintTop_toBottomOf="@id/signInNowButton"/>
         <TextView
-            android:id="@+id/signinPasswordTextView"
+            android:id="@+id/signInToSignUpLinkTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/signinEmailEditText"
-            android:layout_below="@+id/signinUsernameEditText"
-            android:layout_marginTop="-169dp"
-            android:text="@string/password"
-            android:layout_alignLeft="@+id/signinEmailEditText" />
+            android:text="@string/sign_up_title"
+            app:layout_constraintTop_toBottomOf="@id/signInNowButton"
+            app:layout_constraintLeft_toRightOf="@id/signInToSignUpInstructionTextView"
+            android:autoLink="all"
+            android:textColor="@color/bright_events_green"/>
 
-    </RelativeLayout>
+    </android.support.constraint.ConstraintLayout>
+
+
+</RelativeLayout>
