+++ app/src/main/res/layout/activity_signup.xml
@@ -1,30 +1,37 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <android.support.constraint.ConstraintLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true">
 
         <TextView
             android:id="@+id/signUpTitleTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_alignParentLeft="true"
             android:layout_alignParentStart="true"
             android:layout_alignParentTop="true"
+            android:layout_marginLeft="23dp"
             android:layout_marginStart="23dp"
             android:layout_marginTop="18dp"
             android:text="@string/sign_up_title"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="23dp"
-            android:textSize="30sp"/>
+            android:textSize="30sp" />
 
         <EditText
             android:id="@+id/signupEmailEditText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
+            android:layout_alignParentTop="true"
+            android:layout_centerHorizontal="true"
             android:ems="10"
+            android:hint="@string/email"
             android:inputType="textEmailAddress"
-            android:text="@string/email" />
+            app:layout_constraintTop_toBottomOf="@id/signUpTitleTextView" />
 
         <EditText
             android:id="@+id/signupUsernameEditText"
@@ -32,68 +39,63 @@
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginBottom="169dp"
+            android:layout_marginBottom="222dp"
             android:ems="10"
+            android:hint="@string/username"
             android:inputType="textPersonName"
-            android:text="@string/username" />
+            app:layout_constraintTop_toBottomOf="@id/signupEmailEditText" />
 
         <EditText
-            android:id="@+id/confirmPasswordEditText3"
+            android:id="@+id/signupConfirmPasswordEditText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginBottom="40dp"
+            android:layout_marginBottom="72dp"
             android:ems="10"
+            android:hint="@string/confirm_password"
             android:inputType="textPassword"
-            android:text="@string/password"/>
+            app:layout_constraintTop_toBottomOf="@id/signupUsernameEditText" />
 
         <EditText
-            android:id="@+id/PasswordEditText"
+            android:id="@+id/signupPasswordEditText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginBottom="104dp"
+            android:layout_marginBottom="146dp"
             android:ems="10"
+            android:hint="@string/password"
             android:inputType="textPassword"
-            android:text="@string/password"/>
+            app:layout_constraintTop_toBottomOf="@id/signupConfirmPasswordEditText" />
 
-        <TextView
-            android:id="@+id/signupEmailTextView"
+        <Button
+            android:id="@+id/signUpNowButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_above="@+id/signupEmailEditText"
-            android:layout_alignStart="@+id/signupEmailEditText"
-            android:text="@string/email_title"
-            android:layout_alignLeft="@+id/signupEmailEditText" />
+            android:layout_alignParentBottom="true"
+            android:layout_centerHorizontal="true"
+            android:layout_marginBottom="16dp"
+            android:text="@string/sign_up_now_button"
+            app:layout_constraintTop_toBottomOf="@id/signupPasswordEditText" />
 
         <TextView
-            android:id="@+id/signUpUsernameTextView"
+            android:id="@+id/signupToSigninInstructionTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/signupEmailEditText"
-            android:layout_below="@+id/signupEmailEditText"
-            android:text="@string/username"
-            android:layout_alignLeft="@+id/signupEmailEditText" />
+            android:text="@string/signup_to_signin_instruction_textview"
+            app:layout_constraintTop_toBottomOf="@id/signUpNowButton"/>
 
         <TextView
-            android:id="@+id/signupPasswordTextView"
+            android:id="@+id/signUpToSignInLinkTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/signupEmailEditText"
-            android:layout_below="@+id/signupUsernameEditText"
-            android:layout_marginTop="-169dp"
-            android:text="@string/password"
-            android:layout_alignLeft="@+id/signupEmailEditText" />
+            android:text="@string/sign_in_title"
+            app:layout_constraintTop_toBottomOf="@id/signUpNowButton"
+            app:layout_constraintLeft_toRightOf="@id/signupToSigninInstructionTextView"
+            android:textColor="@color/bright_events_green"/>
 
-        <TextView
-            android:id="@+id/signupComfirmPasswordTextView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_above="@+id/confirmPasswordEditText3"
-            android:layout_alignStart="@+id/signupEmailEditText"
-            android:text="@string/confirm_password"
-            android:layout_alignLeft="@+id/signupEmailEditText" />
+
+    </android.support.constraint.ConstraintLayout>
 
 </RelativeLayout>
