+++ app/src/main/res/values/strings.xml
@@ -18,4 +18,7 @@
     <string name="edit_event">Edit</string>
     <string name="delete">Delete</string>
     <string name="rsvp">RSVP</string>
+    <string name="sign_up_now_button">Sign Up!</string>
+    <string name="signin_to_signup_textview">Don\'t have an account?</string>
+    <string name="signup_to_signin_instruction_textview">Already have an account?</string>
 </resources>
